File name checking and removal moved to utility file and added into model and mask
