+++ src/TrainingSrc/mask_test.py
@@ -1,21 +1,23 @@
 import os
 import numpy as np
 from PIL import Image
+from Utilities.custom_utils import get_image_fnames, remove_unmatched_fnames
+
 import pytest
 
-def get_image_fnames(mask_fnames):
-    image_fnames = [mask.replace('_mask', '') for mask in mask_fnames]
-    return image_fnames
+# def get_image_fnames(mask_fnames):
+#     image_fnames = [mask.replace('_mask', '') for mask in mask_fnames]
+#     return image_fnames
 
-def remove_unmatched_fnames(image_fnames, mask_fnames, all_image_fnames):
-    missing_images = list(set(image_fnames).difference(set(all_image_fnames)))
-    if len(missing_images)>0 :
-        print('File Name Error: files removed')
-        print(missing_images)
-        for m_image in missing_images:
-            image_fnames.remove(m_image)
-            mask_fnames.remove(m_image.replace('.', '_mask.'))
-    return image_fnames, mask_fnames
+# def remove_unmatched_fnames(image_fnames, mask_fnames, all_image_fnames):
+#     missing_images = list(set(image_fnames).difference(set(all_image_fnames)))
+#     if len(missing_images)>0 :
+#         print('File Name Error: files removed')
+#         print(missing_images)
+#         for m_image in missing_images:
+#             image_fnames.remove(m_image)
+#             mask_fnames.remove(m_image.replace('.', '_mask.'))
+#     return image_fnames, mask_fnames
 
 mask_fnames = list(sorted(os.listdir("Data/PNGMasks")))
 image_fnames = get_image_fnames(mask_fnames)
